requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
boto3==1.38.35
botocore==1.38.35
catalogue==2.0.10
cdx_toolkit==0.9.37
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
murmurhash==1.0.13
numpy==2.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
preshed==3.0.10
propcache==0.3.2
psycopg2-binary==2.9.10
py4j==0.10.9.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
warcio==1.7.5
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
